+++ src/exec/utils.c
@@ -6,7 +6,7 @@
 /*   By: rgallien <rgallien@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/03 20:42:34 by rgallien          #+#    #+#             */
-/*   Updated: 2024/10/14 16:44:28 by rgallien         ###   ########.fr       */
+/*   Updated: 2024/10/15 11:38:20 by rgallien         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ void	my_mlx_pixel_put(t_img *img, int x, int y, int color)
 	*(unsigned int *)dst = color;
 }
 
-double	to_radiant(unsigned long	number)
+double	to_radiant(double	number)
 {
 	return (number * (M_PI / 180));
 }
